Edited views and use cases to prevent coupling

+++ src/main/java/usecase/deleteproject/DeleteProjectInteractor.java
@@ -1,20 +1,28 @@
 package usecase.deleteproject;
 
 import dataaccess.DataAccessConfig;
+import dataaccess.ILoginUserDetails;
 import dataaccess.IProjectRepository;
 
 public class DeleteProjectInteractor implements DeleteProjectInputBoundary{
     private final IProjectRepository projectRepository;
-    private final DeleteProjectPresenter deleteProjectPresenter;
+    private final ILoginUserDetails loginUserDetails = DataAccessConfig.getLoginUserDetails();
+    private final DeleteProjectOutputBoundary deleteProjectPresenter;
 
-    public DeleteProjectInteractor(DeleteProjectPresenter deleteProjectPresenter){
+    public DeleteProjectInteractor(DeleteProjectOutputBoundary deleteProjectPresenter){
         this.projectRepository = DataAccessConfig.getProjectRepository();
         this.deleteProjectPresenter = deleteProjectPresenter;
     }
 
     @Override
     public void deleteProject(DeleteProjectInputData inputData) {
+        int userId = loginUserDetails.getUserId();
         int projectId = inputData.getProjectId();
+        if (projectRepository.getOwnerId(projectId) != userId) {
+            deleteProjectPresenter.prepareFailView("Only the owner of the project can delete it.");
+            return;
+        }
+
         projectRepository.deleteProject(projectId);
         deleteProjectPresenter.prepareSuccessView(new DeleteProjectOutputData(projectId));
     }
